% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{BelgiumMaps.OpenStreetMap-package}
\alias{BelgiumMaps.OpenStreetMap}
\alias{BelgiumMaps.OpenStreetMap-package}
\title{Maps of Belgium extracted from OpenStreetMap}
\description{
Maps of Belgium extracted from OpenStreetMap.
Contains places, points, railways, roads, waterways and landused in Belgium
}
\examples{
## Other datasets
data(BE_OSM_PLACES)
data(BE_OSM_POINTS)
data(BE_OSM_RAILWAYS)
data(BE_OSM_ROADS)
data(BE_OSM_WATERWAYS)
data(BE_OSM_LANDUSE)
str(as.data.frame(BE_OSM_LANDUSE))
plot(BE_OSM_RAILWAYS)
plot(BE_OSM_WATERWAYS)
plot(subset(BE_OSM_ROADS, type \%in\% c("motorway", "primary")))
str(as.data.frame(BE_OSM_ROADS))
}

